requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-autoslug==1.9.9
django-cities-light==3.10.1
idna==3.7
progressbar2==4.4.2
python-dotenv==1.0.1
python-utils==3.8.2
pytz==2024.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
